Move conversion logic to ConversionButton component
